Task_04/study_notes_agent/requirements.txt
streamlit
pypdf
openai-agents
python-dotenv